refactor: improve functionality
Refactor all roman numeral conversion functions so that they return an object containing a roman numeral equal and a number. The number is NOT equal to the roman numeral value. The number is the user input minus the value of the roman numeral computed within that function.

For example, if the user input is 4567, the object returned by the findThousands function will include the roman numeral MMMM and the number 567.